+++ common/src/main/java/dev/lavalink/youtube/YoutubeSourceOptions.java
@@ -1,9 +1,13 @@
 package dev.lavalink.youtube;
 
+import org.jetbrains.annotations.Nullable;
+
 public class YoutubeSourceOptions {
     private boolean allowSearch = true;
     private boolean allowDirectVideoIds = true;
     private boolean allowDirectPlaylistIds = true;
+    private String remoteCipherUrl;
+    private String remoteCipherPassword;
 
     public boolean isAllowSearch() {
         return allowSearch;
@@ -31,4 +35,19 @@ public class YoutubeSourceOptions {
         this.allowDirectPlaylistIds = allowDirectPlaylistIds;
         return this;
     }
+
+    public String getRemoteCipherUrl() {
+        return remoteCipherUrl;
+    }
+
+    public YoutubeSourceOptions setRemoteCipherUrl(String remoteCipherUrl, @Nullable String remoteCipherPassword) {
+        this.remoteCipherUrl = remoteCipherUrl;
+        this.remoteCipherPassword = remoteCipherPassword;
+        return this;
+    }
+
+    public String getRemoteCipherPassword() {
+        return remoteCipherPassword;
+    }
+
 }
