+++ plugin/src/main/java/dev/lavalink/youtube/plugin/YoutubeRemoteCipherConfig.java
@@ -1,0 +1,23 @@
+package dev.lavalink.youtube.plugin;
+
+public class YoutubeRemoteCipherConfig {
+    private String url;
+    private String password;
+
+    public String getUrl() {
+        return url;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setUrl(String url) {
+        this.url = url;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+}
